+++ packages/blade/src/components/Box/BaseBox/types/propsTypes.ts
@@ -207,7 +207,8 @@ type StyledPropsBlade = Partial<
         | 'gridColumnStart'
         | 'gridColumnEnd'
         | 'gridArea'
-      >,
+      > &
+      Pick<LayoutProps, 'display'>,
     '__brand__'
   >
 >;