+++ packages/blade/src/components/Icons/_Svg/Svg/Svg.web.tsx
@@ -5,10 +5,12 @@ import { metaAttribute, MetaConstants } from '~utils/metaAttribute';
 import { useStyledProps } from '~components/Box/styledProps';
 import { makeAccessible } from '~utils/makeAccessible';
 
-const StyledSvg = styled.svg<SvgProps>((props) => {
-  const styledPropsCSSObject = useStyledProps(props);
-  return styledPropsCSSObject;
-});
+const StyledSvg = styled.svg<SvgProps & { styledDisplay: SvgProps['display'] }>(
+  ({ styledDisplay, ...props }) => {
+    const styledPropsCSSObject = useStyledProps({ ...props, display: styledDisplay });
+    return styledPropsCSSObject;
+  },
+);
 
 const Svg = ({
   children,
@@ -17,6 +19,7 @@ const Svg = ({
   width,
   fill,
   testID,
+  display,
   ...styledProps
 }: SvgProps): ReactElement => {
   return (
@@ -27,6 +30,8 @@ const Svg = ({
       viewBox={viewBox}
       width={width}
       fill={fill}
+      // svg has its own display prop, which conflicts with our styled display prop
+      styledDisplay={display}
       {...styledProps}
     >
       {children}